returns 404 when city_id is not linked to any Citi object for delete method

+++ api/v1/views/cities.py
@@ -42,6 +42,8 @@ def deleteCity(city_id):
     """ Deletes a City object """
     try:
         city = storage.get("City", city_id)
+        if city is None:
+            abort (404)
         storage.delete(city)
         storage.save()
         return (jsonify({}), 200)
